supplier integrity

+++ Core API/Team7/Controllers/SupplierController.cs
@@ -98,6 +98,9 @@ namespace Team7.Controllers
 
             Supplier supplier = await this._supplierRepo.GetSupplierIdAsync(id);
 
+            if (supplier.Orders != null)
+                return Conflict(new { supplier = supplier });
+
             _saleItemRepo.Delete(supplier);
             await _supplierRepo.SaveChangesAsync();
